different GBPs

+++ CNN/GBP_Denoising/prepare_data.py
@@ -29,7 +29,7 @@ def Map(tfOp, ph, images, sess):
 
     batch_size = 256
     num_examples = images.shape[0]
-    num_batches = (num_examples - 1) / batch_size + 1
+    num_batches = int((num_examples - 1) / batch_size) + 1
     counter = 0
     result = None
